+++ services/core-banking/src/main/java/com/bank/deposit/dto/request/AccountLimitUpdateRequest.java
@@ -3,7 +3,7 @@ package com.bank.deposit.dto.request;
 import java.math.BigDecimal;
 
 public record AccountLimitUpdateRequest(
-        BigDecimal dailyWithdrawLimit,
+        Long dailyWithdrawLimit,
         Integer dailyWithdrawCountLimit,
-        BigDecimal atmWithdrawLimit
+        Long atmWithdrawLimit
 ) {}
