+++ services/core-banking/src/main/java/com/bank/deposit/dto/request/SubscriptionProductRequest.java
@@ -6,8 +6,8 @@ import jakarta.validation.constraints.Positive;
 import java.math.BigDecimal;
 
 public record SubscriptionProductRequest(
-        @NotNull @Positive BigDecimal monthlyPaymentAmount,
-        BigDecimal minMonthlyPayment,
-        BigDecimal maxMonthlyPayment,
-        BigDecimal maxRecognizedPaymentAmount
+        @NotNull @Positive Long monthlyPaymentAmount,
+        Long minMonthlyPayment,
+        Long maxMonthlyPayment,
+        Long maxRecognizedPaymentAmount
 ) {}
