+++ services/core-banking/src/main/java/com/bank/deposit/dto/response/CashFlowSummary.java
@@ -1,10 +1,14 @@
 package com.bank.deposit.dto.response;
 
-import java.math.BigDecimal;
-
+/**
+ * 현금흐름 요약. 네 값 모두 원 단위 정수다(데이터 사전: 금액 = BIGINT 원 단위).
+ *
+ * <p>{@code estimatedSavingsAmount} 는 순현금흐름을 개월 수로 나눈 값이라 계산 중에는
+ * BigDecimal 이지만, scale 0 (RoundingMode.DOWN) 으로 절사해 담으므로 정수다.
+ */
 public record CashFlowSummary(
-        BigDecimal totalInflow,
-        BigDecimal totalOutflow,
-        BigDecimal netCashFlow,
-        BigDecimal estimatedSavingsAmount
+        Long totalInflow,
+        Long totalOutflow,
+        Long netCashFlow,
+        Long estimatedSavingsAmount
 ) {}
