+++ services/core-banking/src/main/java/com/bank/deposit/dto/response/RecommendedProduct.java
@@ -8,8 +8,8 @@ public record RecommendedProduct(
         String productType,
         BigDecimal baseInterestRate,
         BigDecimal bestRate,
-        BigDecimal minJoinAmount,
-        BigDecimal maxJoinAmount,
+        Long minJoinAmount,
+        Long maxJoinAmount,
         Integer minPeriodMonth,
         Integer maxPeriodMonth,
         String reason
